account permissions almost done edite and delete admin accounts

+++ PhysicsLabsDB/Accounts/frmAccounts.cs
@@ -16,6 +16,8 @@ namespace PhysicsLabsDB.Accounts
     {
         physics_dbEntities db = new physics_dbEntities();
 
+        string selectedUserToEdit = null;
+
         Status CurrentStatus = Status.Reset;
 
         public frmAccounts()
@@ -26,11 +28,13 @@ namespace PhysicsLabsDB.Accounts
         public void ClearControls()
         {
             txtuserName.Text = string.Empty;
-			txtPassword.Text = string.Empty;
-			txtConfirmPassword.Text = string.Empty;
+            txtPassword.Text = string.Empty;
+            txtConfirmPassword.Text = string.Empty;
 
-			cmbEmployee.SelectedItem = null;
+            cmbEmployee.SelectedItem = null;
             cmbAccountType.SelectedItem = null;
+
+            selectedUserToEdit = null;
         }
 
         public void ControlStatus(bool status)
@@ -38,9 +42,9 @@ namespace PhysicsLabsDB.Accounts
             txtuserName.Enabled =
             cmbEmployee.Enabled =
             cmbAccountType.Enabled =
-			txtConfirmPassword.Enabled =
-			txtPassword.Enabled = status;
-			grdVwUsers.Enabled = status;
+            txtConfirmPassword.Enabled =
+            txtPassword.Enabled = status;
+            grdVwUsers.Enabled = status;
         }
 
         private enum Status
@@ -51,72 +55,72 @@ namespace PhysicsLabsDB.Accounts
             Save
         };
 
-		private void ToolStripButtonStatus(Status status)
-		{
-			CurrentStatus = status;
-
-			if (status == Status.New)
-			{
-				newToolStripButton.Enabled = false;
-				openToolStripButton.Enabled = false;
-				saveToolStripButton.Enabled = true;
-				delToolStripButton.Enabled = false;
-				resetToolStripButton.Enabled = true;
-			}
-			if (status == Status.Edit)
-			{
-				newToolStripButton.Enabled = false;
-				openToolStripButton.Enabled = false;
-				saveToolStripButton.Enabled = true;
-				delToolStripButton.Enabled = true;
-				resetToolStripButton.Enabled = true;
-			}
-			if (status == Status.Reset)
-			{
-				newToolStripButton.Enabled = true;
-				openToolStripButton.Enabled = true;
-				saveToolStripButton.Enabled = false;
-				delToolStripButton.Enabled = false;
-				resetToolStripButton.Enabled = false;
-			}
-			if (status == Status.Save)
-			{
-				newToolStripButton.Enabled = true;
-				openToolStripButton.Enabled = true;
-				saveToolStripButton.Enabled = false;
-				delToolStripButton.Enabled = false;
-				resetToolStripButton.Enabled = true;
-			}
-		}
-
-		private void frmAccounts_Load(object sender, EventArgs e)
+        private void ToolStripButtonStatus(Status status)
         {
-			grdVwUsers.AutoGenerateColumns = false;
+            CurrentStatus = status;
+
+            if (status == Status.New)
+            {
+                newToolStripButton.Enabled = false;
+                openToolStripButton.Enabled = false;
+                saveToolStripButton.Enabled = true;
+                delToolStripButton.Enabled = false;
+                resetToolStripButton.Enabled = true;
+            }
+            if (status == Status.Edit)
+            {
+                newToolStripButton.Enabled = false;
+                openToolStripButton.Enabled = false;
+                saveToolStripButton.Enabled = true;
+                delToolStripButton.Enabled = true;
+                resetToolStripButton.Enabled = true;
+            }
+            if (status == Status.Reset)
+            {
+                newToolStripButton.Enabled = true;
+                openToolStripButton.Enabled = true;
+                saveToolStripButton.Enabled = false;
+                delToolStripButton.Enabled = false;
+                resetToolStripButton.Enabled = false;
+            }
+            if (status == Status.Save)
+            {
+                newToolStripButton.Enabled = true;
+                openToolStripButton.Enabled = true;
+                saveToolStripButton.Enabled = false;
+                delToolStripButton.Enabled = false;
+                resetToolStripButton.Enabled = true;
+            }
+        }
+
+        private void frmAccounts_Load(object sender, EventArgs e)
+        {
+            grdVwUsers.AutoGenerateColumns = false;
 
             cmbEmployee.DataSource = db.respons.ToList();
-			cmbEmployee.DisplayMember = "name";
-			cmbEmployee.ValueMember = "ID";
+            cmbEmployee.DisplayMember = "name";
+            cmbEmployee.ValueMember = "ID";
 
-			cmbAccountType.DataSource = db.account_types.ToList();
-            
+            cmbAccountType.DataSource = db.account_types.ToList();
+
             //cmbAccountType.DataSource = db.account_types.Select(u => u.AccountType).ToList();
-			cmbAccountType.DisplayMember = "AccountType";
-			cmbAccountType.ValueMember = "ID";
-			Search();
+            cmbAccountType.DisplayMember = "AccountType";
+            cmbAccountType.ValueMember = "ID";
+            Search();
 
-			ClearControls();
-			ControlStatus(false);
-			ToolStripButtonStatus(Status.Reset);
-		}
+            ClearControls();
+            ControlStatus(false);
+            ToolStripButtonStatus(Status.Reset);
+        }
 
         private void newToolStripButton_Click(object sender, EventArgs e)
         {
-			ClearControls();
-			ControlStatus(true);
-			grdVwUsers.Enabled = false;
-			ToolStripButtonStatus(Status.New);
-			// todo
-		}
+            ClearControls();
+            ControlStatus(true);
+            grdVwUsers.Enabled = false;
+            ToolStripButtonStatus(Status.New);
+            // todo
+        }
 
 
         private void saveToolStripButton_Click(object sender, EventArgs e)
@@ -124,11 +128,11 @@ namespace PhysicsLabsDB.Accounts
             string error = string.Empty;
             if (txtuserName.Text == string.Empty)
                 error += "إدخل اسم المستخدم\n";
-			if (txtPassword.Text == string.Empty)
-				error += "إدخل كلمة السر\n";
-			if (txtConfirmPassword.Text == string.Empty)
-				error += "إدخل تأكيد كلمة السر\n";
-			if (cmbAccountType.SelectedItem == null)
+            if (txtPassword.Text == string.Empty)
+                error += "إدخل كلمة السر\n";
+            if (txtConfirmPassword.Text == string.Empty)
+                error += "إدخل تأكيد كلمة السر\n";
+            if (cmbAccountType.SelectedItem == null)
                 error += "اختر نوع الحساب\n";
             if (cmbEmployee.SelectedItem == null)
                 error += "اختر الموظف\n";
@@ -142,67 +146,72 @@ namespace PhysicsLabsDB.Accounts
             {
                 try
                 {
-					if (CurrentStatus == Status.New)
-					{
-						if (CheckUserFound(txtuserName.Text))
-						{
-							MessageBox.Show("اسم المستخدم موجود مسبقاً", "خطأ", MessageBoxButtons.OK, MessageBoxIcon.Information);
-							return;
-						}
-						if (txtPassword.Text != txtConfirmPassword.Text)
-						{
-							MessageBox.Show("كلمة المرور غير متطابقة", "خطأ", MessageBoxButtons.OK, MessageBoxIcon.Information);
-							return;
-						}
-						else
-						{
-							var account = new account()
-							{
-								UserName = txtuserName.Text,
-								AccountType = Convert.ToInt32(cmbAccountType.SelectedValue),
-								UserPassword = txtPassword.Text,
-								Respon = Convert.ToInt32(cmbEmployee.SelectedValue),
-							};
+                    if (CurrentStatus == Status.New)
+                    {
+                        if (CheckUserFound(txtuserName.Text))
+                        {
+                            MessageBox.Show("اسم المستخدم موجود مسبقاً", "خطأ", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                            return;
+                        }
+                        if (txtPassword.Text != txtConfirmPassword.Text)
+                        {
+                            MessageBox.Show("كلمة المرور غير متطابقة", "خطأ", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                            return;
+                        }
+                        var account = new account()
+                        {
+                            UserName = txtuserName.Text,
+                            AccountType = Convert.ToInt32(cmbAccountType.SelectedValue),
+                            UserPassword = txtPassword.Text,
+                            Respon = Convert.ToInt32(cmbEmployee.SelectedValue),
+                        };
 
-							db.accounts.Add(account);
-							db.SaveChanges();
-						}
-						MessageBox.Show("تم الحفظ", "تم", MessageBoxButtons.OK, MessageBoxIcon.Information);
-						
+                        db.accounts.Add(account);
+                        db.SaveChanges();
+                        MessageBox.Show("تم الحفظ", "تم", MessageBoxButtons.OK, MessageBoxIcon.Information);
 
-						//ClearControls();
-						ControlStatus(false);
-						ToolStripButtonStatus(Status.Save);
-						Search();
-					}
+
+                        //ClearControls();
+                        ControlStatus(false);
+                        ToolStripButtonStatus(Status.Save);
+                        Search();
+                    }
                     if (CurrentStatus == Status.Edit)
                     {
-						if (CheckUserFound(txtuserName.Text))
-						{
-							MessageBox.Show("اسم المستخدم موجود مسبقاً", "خطأ", MessageBoxButtons.OK, MessageBoxIcon.Information);
-							return;
-						}
-						if (txtPassword.Text != txtConfirmPassword.Text)
-						{
-							MessageBox.Show("كلمة المرور غير متطابقة", "خطأ", MessageBoxButtons.OK, MessageBoxIcon.Information);
-							return;
-						}
-						else
-						{
-							var account = db.accounts.FirstOrDefault(u => u.UserName == txtuserName.Text);
-							account.UserName = txtuserName.Text;
-							account.AccountType = Convert.ToInt32(cmbAccountType.SelectedValue);
-							account.UserPassword = txtuserName.Text;
-							account.Respon = Convert.ToInt32(cmbEmployee.Text);
-							db.SaveChanges();
+                        if (selectedUserToEdit != txtuserName.Text && CheckUserFound(txtuserName.Text))
+                        {
+                            MessageBox.Show("اسم المستخدم موجود مسبقاً", "خطأ", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                            return;
+                        }
+                        if (txtPassword.Text != txtConfirmPassword.Text)
+                        {
+                            MessageBox.Show("كلمة المرور غير متطابقة", "خطأ", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                            return;
+                        }
+                        var account = db.accounts.FirstOrDefault(u => u.UserName == selectedUserToEdit);
+                        
+                        if (account.account_types.AccountType == "admin")
+                        {
+                            var adminAccountsCount = db.accounts.Where(u => u.AccountType == account.AccountType).Count();
+                            if (adminAccountsCount == 1 && Convert.ToInt32(cmbAccountType.SelectedValue) != account.AccountType)
+                            {
+                                MessageBox.Show("لا يمكن تعديل صلاحيات هذا الحساب", "خطأ", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                                return;
+                            }
+                        }
 
-							MessageBox.Show("تم التعديل", "تم", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                        account.UserName = txtuserName.Text;
+                        account.AccountType = Convert.ToInt32(cmbAccountType.SelectedValue);
+                        account.UserPassword = txtPassword.Text;
+                        account.Respon = Convert.ToInt32(cmbEmployee.SelectedValue);
+                        db.SaveChanges();
 
-							//ClearControls();
-							ControlStatus(false);
-							ToolStripButtonStatus(Status.Save);
-							Search();
-						}
+                        MessageBox.Show("تم التعديل", "تم", MessageBoxButtons.OK, MessageBoxIcon.Information);
+
+                        //ClearControls();
+                        ControlStatus(false);
+                        ToolStripButtonStatus(Status.Save);
+                        Search();
                     }
                 }
                 catch (Exception ex)
@@ -217,10 +226,19 @@ namespace PhysicsLabsDB.Accounts
         {
             //todo
             //..
-			//if()
+            //if()
             try
             {
                 var account = db.accounts.FirstOrDefault(u => u.UserName == txtuserName.Text);
+                if(account.account_types.AccountType == "admin")
+                {
+                    var adminAccountsCount = db.accounts.Where(u => u.AccountType == account.AccountType).Count();
+                    if (adminAccountsCount == 1)
+                    {
+                        MessageBox.Show("لا يمكن حذف هذا الحساب", "خطأ", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                        return;
+                    }
+                }
                 db.accounts.Remove(account);
                 db.SaveChanges();
 
@@ -229,8 +247,8 @@ namespace PhysicsLabsDB.Accounts
                 ClearControls();
                 ControlStatus(false);
                 ToolStripButtonStatus(Status.Reset);
-				Search();
-			}
+                Search();
+            }
             catch (Exception ex)
             {
                 MessageBox.Show(ex.ToString(), "خطأ", MessageBoxButtons.OK, MessageBoxIcon.Error);
@@ -238,44 +256,44 @@ namespace PhysicsLabsDB.Accounts
         }
 
 
-		public void Search()
-		{
-			var result = db.accounts.Select(u => new
-			{
-				UserName = u.UserName,
-				UserPassword = u.UserPassword,
-				ID = u.ID,
-				Respon = u.respon1.name,
-				AccountType = u.account_types.AccountType
-			}).ToList();
-			grdVwUsers.DataSource = result;
-			//object username;
+        public void Search()
+        {
+            var result = db.accounts.Select(u => new
+            {
+                UserName = u.UserName,
+                UserPassword = u.UserPassword,
+                ID = u.ID,
+                Respon = u.respon1.name,
+                AccountType = u.account_types.AccountType
+            }).ToList();
+            grdVwUsers.DataSource = result;
+            //object username;
 
-			//// must add reference to System.Configuration  from assemblies
-			//SqlConnection connection = new SqlConnection(System.Configuration.ConfigurationManager.ConnectionStrings["ConnectionStringSQL"].ConnectionString);
-			//SqlDataAdapter data_adabter = new SqlDataAdapter();
-			//DataTable result = new DataTable();
+            //// must add reference to System.Configuration  from assemblies
+            //SqlConnection connection = new SqlConnection(System.Configuration.ConfigurationManager.ConnectionStrings["ConnectionStringSQL"].ConnectionString);
+            //SqlDataAdapter data_adabter = new SqlDataAdapter();
+            //DataTable result = new DataTable();
 
-			//string query = @"select UserName,
-			//         	   Respon,
-			//         	   AccountType
-			//         from account 
-			//         	   where (@username is null or UserName like @username + '%');";
+            //string query = @"select UserName,
+            //         	   Respon,
+            //         	   AccountType
+            //         from account 
+            //         	   where (@username is null or UserName like @username + '%');";
 
-			//connection.Open();
-			//data_adabter = new SqlDataAdapter(query, connection);
-			//data_adabter.SelectCommand.Parameters.AddWithValue("@username", username);
-			//data_adabter.Fill(result);
-			//connection.Close();
+            //connection.Open();
+            //data_adabter = new SqlDataAdapter(query, connection);
+            //data_adabter.SelectCommand.Parameters.AddWithValue("@username", username);
+            //data_adabter.Fill(result);
+            //connection.Close();
 
-			grdVwUsers.Columns["colUserName"].DataPropertyName = "UserName";
-			grdVwUsers.Columns["colAccountType"].DataPropertyName = "AccountType";
-			grdVwUsers.Columns["colRespon"].DataPropertyName = "Respon";
-			//grdVwUsers.DataSource = result;
-		}
+            grdVwUsers.Columns["colUserName"].DataPropertyName = "UserName";
+            grdVwUsers.Columns["colAccountType"].DataPropertyName = "AccountType";
+            grdVwUsers.Columns["colRespon"].DataPropertyName = "Respon";
+            //grdVwUsers.DataSource = result;
+        }
 
 
-		private void resetToolStripButton_Click(object sender, EventArgs e)
+        private void resetToolStripButton_Click(object sender, EventArgs e)
         {
             //cmbEmployee.DataSource = db.respons.Select(u => u.name).ToList();
             //cmbLab.DataSource = db.labs.Select(u => u.lab_name).ToList();
@@ -287,40 +305,36 @@ namespace PhysicsLabsDB.Accounts
 
         private void grdVwSearch_CellClick(object sender, DataGridViewCellEventArgs e)
         {
-			//if(CurrentStatus = Status.Edit)
-            var rowIndex = grdVwUsers.CurrentRow.Index; //SelectedCells[0].RowIndex;
+            var rowIndex = grdVwUsers.CurrentRow.Index;
             ClearControls();
             ControlStatus(true);
-			//txtuserName.Enabled = false;
-            //DataGridViewRow selectedRow = grdVwSearch.Rows[rowIndex];
             txtuserName.Text = grdVwUsers.Rows[rowIndex].Cells[0].Value.ToString();
-			cmbEmployee.SelectedText = grdVwUsers.Rows[rowIndex].Cells[1].Value.ToString();
-			cmbAccountType.SelectedText = grdVwUsers.Rows[rowIndex].Cells[2].Value.ToString();
+            selectedUserToEdit = txtuserName.Text;
+            cmbAccountType.Text = grdVwUsers.Rows[rowIndex].Cells[1].Value.ToString();
+            cmbEmployee.Text = grdVwUsers.Rows[rowIndex].Cells[2].Value.ToString();
+            txtPassword.Text = txtConfirmPassword.Text 
+                = db.accounts.Where(u => u.UserName == txtuserName.Text)
+                .Select(u => u.UserPassword).ToList()[0];
         }
 
-       
-		private void lblEmployee_Click(object sender, EventArgs e)
-		{
 
-		}
+        private void lblEmployee_Click(object sender, EventArgs e)
+        {
 
-		private void openToolStripButton_Click(object sender, EventArgs e)
-		{
-			ClearControls();
-			ControlStatus(false);
-			ToolStripButtonStatus(Status.Edit);
-			//btnDisplayBarcode.Enabled = false;
-			grdVwUsers.Enabled = true;
-			Search();
-		}
+        }
 
-		private bool CheckUserFound(string user)
-		{
-			var users = db.accounts.Select(u => u.UserName == user).ToList();
-			if (users != null)
-				return true;
-			else
-				return false;
-		}
-	}
+        private void openToolStripButton_Click(object sender, EventArgs e)
+        {
+            ClearControls();
+            ControlStatus(false);
+            ToolStripButtonStatus(Status.Edit);
+            grdVwUsers.Enabled = true;
+            Search();
+        }
+
+        private bool CheckUserFound(string user)
+        {
+            return db.accounts.Any(u => u.UserName == user);
+        }
+    }
 }
